payloads can be longer than 255 bytes

+++ include/snmp_elem.h
@@ -18,7 +18,7 @@ public:
 
 	virtual uint8_t get_size() const { return len; }
 
-	virtual std::pair<uint8_t *, uint8_t> get_payload() const = 0;
+	virtual std::pair<uint8_t *, size_t> get_payload() const = 0;
 };
 
 //---
@@ -37,7 +37,7 @@ public:
 	explicit snmp_integer(const snmp_integer_type type, const uint64_t v);
 	virtual ~snmp_integer();
 
-	std::pair<uint8_t *, uint8_t> get_payload() const override;
+	std::pair<uint8_t *, size_t> get_payload() const override;
 };
 
 //---
@@ -55,7 +55,7 @@ public:
 
 	uint8_t get_size() const override;
 
-	std::pair<uint8_t *, uint8_t> get_payload() const override;
+	std::pair<uint8_t *, size_t> get_payload() const override;
 };
 
 //---
@@ -66,7 +66,7 @@ public:
 	snmp_null();
 	virtual ~snmp_null();
 
-	std::pair<uint8_t *, uint8_t> get_payload() const override;
+	std::pair<uint8_t *, size_t> get_payload() const override;
 };
 
 //---
@@ -80,7 +80,7 @@ public:
 	explicit snmp_octet_string(const uint8_t *const v, const int len);
 	virtual ~snmp_octet_string();
 
-	std::pair<uint8_t *, uint8_t> get_payload() const override;
+	std::pair<uint8_t *, size_t> get_payload() const override;
 };
 
 //---
@@ -94,7 +94,7 @@ public:
 	explicit snmp_oid(const std::string & oid);
 	virtual ~snmp_oid();
 
-	std::pair<uint8_t *, uint8_t> get_payload() const override;
+	std::pair<uint8_t *, size_t> get_payload() const override;
 };
 
 //---
@@ -108,5 +108,5 @@ public:
 	explicit snmp_pdu(const uint8_t type);
 	virtual ~snmp_pdu();
 
-	std::pair<uint8_t *, uint8_t> get_payload() const override;
+	std::pair<uint8_t *, size_t> get_payload() const override;
 };
